unifying alignment info
