+++ engeom/src/geom3/align3/information.rs
@@ -1,39 +1,25 @@
 //! Tools for asking how well a set of points constrains a 3D alignment, and for choosing a
 //! subset of them that preserves that conditioning.
 //!
-//! # What this is for
-//!
-//! A points-to-surface alignment is a least-squares problem, and every test point contributes one
-//! row `j_i` to its jacobian. The **information matrix** is the sum of the rank-1 contributions of
-//! those rows,
-//!
-//! $$ H = \sum_i w_i j_i j_i^T $$
-//!
-//! and it is the object that answers both questions this module exists for. Its inverse is (up to
-//! a noise scale) the covariance of the fitted parameters, so it says how well each degree of
-//! freedom is pinned down; and because it is a *sum* over points, the contribution of any single
-//! point is a rank-1 update that can be added or removed in closed form.
-//!
-//! Two distinct uses follow:
+//! The analysis itself is dimension-independent and lives in
+//! [`crate::common::align::information::AlignInformation`], whose module documentation explains
+//! the information matrix and the two uses it serves:
 //!
 //! - **Diagnosis.** Is this alignment well conditioned, and if not, along what motion is it free
 //!   to slide? See [`AlignInformation3::marginal_precision`] and
-//!   [`AlignInformation3::weak_directions`].
+//!   [`AlignInformation3::weak_directions`]. In 3D the weak direction of an alignment is rarely
+//!   axis-aligned: points on a cylinder slide along a helix, points on a plane slide in any
+//!   in-plane direction and spin about the normal, and a shallow dish is weakest in a coupled
+//!   tilt-and-shift.
 //! - **Pruning.** Which points are actually carrying the alignment, and what is the smallest
 //!   subset that would constrain it just as well? See [`AlignInformation3::leverage`] and
-//!   [`AlignInformation3::select_d_optimal`].
+//!   [`AlignInformation3::select_d_optimal`]. Pruning is the reason this exists: a simultaneous
+//!   alignment of fifteen or twenty scans is expensive in proportion to the number of alignment
+//!   points, and most of those points are redundant with their neighbors.
 //!
-//! Pruning is the reason this exists. A simultaneous alignment of fifteen or twenty scans is
-//! expensive in proportion to the number of alignment points, and most of those points are
-//! redundant with their neighbors.
-//!
-//! # Why scoring points individually is not enough
-//!
-//! The obvious approach is to score each point on its own and keep the best `n`. That
-//! double-counts redundancy: two points sitting on the same flat patch each score well, but once
-//! either is in the set the other adds almost nothing. Greedy D-optimal selection avoids this for
-//! free, because a candidate is scored by its *marginal* gain against everything already chosen,
-//! and that gain collapses as soon as something equivalent has been picked.
+//! This module supplies the 3D ingredients: the [`AlignDofs`] implementation on [`Dof6`] and the
+//! [`AlignInformation3::from_points`] constructor, which turns points and a [`SurfaceTarget3`]
+//! into jacobian rows. The `geom2::align2::information` module does the same for 2D.
 //!
 //! # A caveat about units
 //!
@@ -48,42 +34,27 @@
 //! `AlignParams3::from_center` arranges). Placing the origin far away inflates the rotation block
 //! and will make rotations look artificially well constrained.
 
+use crate::Point3;
+use crate::common::align::information::{AlignDofs, AlignInformation};
 use crate::geom3::align3::jacobian::point_surf_jacobian;
-use crate::geom3::align3::{AlignParams3, AlignStorage3, Dof6, SurfaceTarget3};
-use crate::na::{DMatrix, DVector, Matrix6};
-use crate::{Point3, Result};
+use crate::geom3::align3::{AlignParams3, Dof6, SurfaceTarget3};
 
-/// The number of alignment parameters (tx, ty, tz, rx, ry, rz).
-const N_PARAMS: usize = 6;
-
-/// The relative size of the regularizing ridge used to start greedy selection. See
-/// [`AlignInformation3::select_d_optimal`] for why one is needed.
-const DEFAULT_RIDGE_REL: f64 = 1e-9;
-
-/// The information content of a set of test points with respect to a 3D alignment.
+/// The information content of a set of test points with respect to a 3D alignment, over the six
+/// parameters `tx`, `ty`, `tz`, `rx`, `ry`, `rz`.
 ///
 /// Built with [`AlignInformation3::from_points`], which projects the points onto a target and
 /// takes the jacobian row of each correspondence, or with [`AlignInformation3::from_rows`] when
 /// the rows have already been computed. See the module documentation for what it is good for.
-#[derive(Clone, Debug)]
-pub struct AlignInformation3 {
-    /// The jacobian row for each point. Locked degrees of freedom are already zero in these,
-    /// because `point_surf_jacobian` zeroes their columns.
-    rows: Vec<AlignStorage3>,
+pub type AlignInformation3 = AlignInformation<Dof6, 6>;
 
-    /// The weight of each point's contribution to the information matrix.
-    weights: Vec<f64>,
-
-    /// Which degrees of freedom are free to move.
-    dof: Dof6,
-
-    /// The indices, within the six parameters, of the degrees of freedom which are free. All of
-    /// the linear algebra happens in this subspace, because `H` is identically zero along a locked
-    /// direction and would not be invertible in the full space.
-    free: Vec<usize>,
-
-    /// The information matrix in the free subspace, `k` by `k` where `k = free.len()`.
-    h: DMatrix<f64>,
+impl AlignDofs<6> for Dof6 {
+    fn free_indices(&self) -> Vec<usize> {
+        [self.tx, self.ty, self.tz, self.rx, self.ry, self.rz]
+            .iter()
+            .enumerate()
+            .filter_map(|(i, &free)| free.then_some(i))
+            .collect()
+    }
 }
 
 impl AlignInformation3 {
@@ -129,297 +100,13 @@ impl AlignInformation3 {
 
         Self::assemble(rows, weights, params.dof)
     }
-
-    /// Builds the information content directly from jacobian rows which have already been
-    /// computed, for callers assembling a problem this module doesn't know how to build (a
-    /// multi-body adjustment, for instance).
-    ///
-    /// Rows are expected to already have their locked columns zeroed, which is what
-    /// `point_surf_jacobian` produces when given a constrained [`Dof6`].
-    ///
-    /// Returns an error if the two slices disagree in length, or if any weight is negative or
-    /// non-finite.
-    pub fn from_rows(rows: Vec<AlignStorage3>, weights: Vec<f64>, dof: Dof6) -> Result<Self> {
-        if rows.len() != weights.len() {
-            return Err(format!(
-                "there are {} jacobian rows but {} weights",
-                rows.len(),
-                weights.len()
-            )
-            .into());
-        }
-
-        if let Some(i) = weights.iter().position(|w| !w.is_finite() || *w < 0.0) {
-            return Err(format!(
-                "weights[{}] is {}, but every weight must be finite and non-negative",
-                i, weights[i]
-            )
-            .into());
-        }
-
-        Ok(Self::assemble(rows, weights, dof))
-    }
-
-    fn assemble(rows: Vec<AlignStorage3>, weights: Vec<f64>, dof: Dof6) -> Self {
-        let free = free_indices(&dof);
-        let k = free.len();
-
-        let mut h = DMatrix::zeros(k, k);
-        for (row, &w) in rows.iter().zip(weights.iter()) {
-            let j = restrict(row, &free);
-            // The rank-1 contribution of this point, `w * j * j^T`. This uses the general rank-1
-            // update rather than the symmetric one, because `syger` writes only the lower
-            // triangle and every consumer here (the inverse, the quadratic forms) needs the whole
-            // matrix populated.
-            h.ger(w, &j, &j, 1.0);
-        }
-
-        Self {
-            rows,
-            weights,
-            dof,
-            free,
-            h,
-        }
-    }
-
-    /// The number of points this was built from.
-    pub fn len(&self) -> usize {
-        self.rows.len()
-    }
-
-    /// Whether there are no points at all.
-    pub fn is_empty(&self) -> bool {
-        self.rows.is_empty()
-    }
-
-    /// The degrees of freedom that were free when this was built.
-    pub fn dof(&self) -> Dof6 {
-        self.dof
-    }
-
-    /// How many degrees of freedom are free, which is the rank the information matrix would have
-    /// if the points constrained the pose completely.
-    pub fn free_dof_count(&self) -> usize {
-        self.free.len()
-    }
-
-    /// The information matrix, expanded back into the full six parameters with zeros in the rows
-    /// and columns of any locked degree of freedom.
-    ///
-    /// Note the unit caveat in the module documentation: the translation block is dimensionless,
-    /// the rotation block carries units of length squared, and the cross terms fall in between.
-    pub fn information(&self) -> Matrix6<f64> {
-        let mut full = Matrix6::zeros();
-        for (a, &i) in self.free.iter().enumerate() {
-            for (b, &j) in self.free.iter().enumerate() {
-                full[(i, j)] = self.h[(a, b)];
-            }
-        }
-        full
-    }
-
-    /// The marginal precision of each degree of freedom, `1 / [H^-1]_kk`, or `None` for a degree
-    /// of freedom that was locked.
-    ///
-    /// This is the precision of a parameter *accounting for its correlation with the others*,
-    /// which is the quantity that matters and is not the same as the diagonal of `H`. The
-    /// distinction is what a perturbation experiment would measure: displace the points along one
-    /// degree of freedom, let the other five re-optimize to absorb as much of it as they can, and
-    /// the residual sum of squares that survives is `delta^2` times this value. A degree of
-    /// freedom that looks well constrained on the diagonal of `H` can still be nearly free if
-    /// another degree of freedom can imitate it.
-    ///
-    /// Larger is better constrained. The units differ between translation and rotation entries.
-    ///
-    /// Returns an error if the points do not constrain every free degree of freedom, since then
-    /// `H` is singular and no finite precision exists. [`AlignInformation3::weak_directions`]
-    /// still works in that case and will show which motions are unconstrained.
-    pub fn marginal_precision(&self) -> Result<[Option<f64>; N_PARAMS]> {
-        let inverse = self.h.clone().try_inverse().ok_or_else(|| {
-            format!(
-                "the information matrix is singular: the points do not constrain all {} free \
-                 degrees of freedom. Use `weak_directions` to see which motions are free.",
-                self.free.len()
-            )
-        })?;
-
-        let mut out = [None; N_PARAMS];
-        for (a, &i) in self.free.iter().enumerate() {
-            out[i] = Some(1.0 / inverse[(a, a)]);
-        }
-        Ok(out)
-    }
-
-    /// The eigen-decomposition of the information matrix, as `(eigenvalue, direction)` pairs
-    /// sorted from least to most constrained.
-    ///
-    /// The first entry is the motion the point set does least to prevent. This is usually more
-    /// informative than the six per-axis numbers, because the weak direction of an alignment is
-    /// rarely axis-aligned: points on a cylinder slide along a helix, points on a plane slide in
-    /// any in-plane direction and spin about the normal, and a shallow dish is weakest in a
-    /// coupled tilt-and-shift. An eigenvalue at or near zero means that motion is unconstrained.
-    ///
-    /// Directions are unit vectors over the six parameters, with zeros in any locked slot. See the
-    /// module documentation for why the relative scaling of the translation and rotation
-    /// components depends on where the local origin was placed.
-    pub fn weak_directions(&self) -> Vec<(f64, AlignStorage3)> {
-        if self.free.is_empty() {
-            return Vec::new();
-        }
-
-        let eigen = self.h.clone().symmetric_eigen();
-        let mut pairs: Vec<(f64, AlignStorage3)> = eigen
-            .eigenvalues
-            .iter()
-            .enumerate()
-            .map(|(a, &value)| {
-                let column = eigen.eigenvectors.column(a);
-                let mut direction = AlignStorage3::zeros();
-                for (b, &i) in self.free.iter().enumerate() {
-                    direction[i] = column[b];
-                }
-                (value, direction)
-            })
-            .collect();
-
-        pairs.sort_by(|a, b| a.0.total_cmp(&b.0));
-        pairs
-    }
-
-    /// The leverage of each point, `w_i * j_i^T H^-1 j_i`.
-    ///
-    /// Leverage is the classical measure of how much an observation influences its own fitted
-    /// value. It is self-normalizing: the values sum to the number of free degrees of freedom, so
-    /// the average is `free_dof_count / len()` and anything well above that is a point the
-    /// alignment leans on.
-    ///
-    /// This scores each point *independently*, which makes it a good diagnostic but a poor basis
-    /// for pruning, since a pair of interchangeable points both score high. Use
-    /// [`AlignInformation3::select_d_optimal`] to choose a subset.
-    ///
-    /// Returns an error if the information matrix is singular, for the same reason as
-    /// [`AlignInformation3::marginal_precision`].
-    pub fn leverage(&self) -> Result<Vec<f64>> {
-        let inverse = self.h.clone().try_inverse().ok_or_else(|| {
-            format!(
-                "the information matrix is singular: the points do not constrain all {} free \
-                 degrees of freedom, so leverage is undefined.",
-                self.free.len()
-            )
-        })?;
-
-        Ok(self
-            .rows
-            .iter()
-            .zip(self.weights.iter())
-            .map(|(row, &w)| {
-                let j = restrict(row, &self.free);
-                w * (j.transpose() * &inverse * &j)[(0, 0)]
-            })
-            .collect())
-    }
-
-    /// Greedily chooses a subset of `count` points which preserves the conditioning of the
-    /// alignment as well as a subset of that size can.
-    ///
-    /// The criterion is D-optimality: maximize the determinant of the information matrix of the
-    /// chosen subset. Adding a point multiplies that determinant by `1 + w_i j_i^T H^-1 j_i`, so
-    /// each round picks the point with the largest such marginal gain and folds it in with a
-    /// Sherman-Morrison rank-1 update of the running inverse. Because a candidate is measured
-    /// against what has already been chosen rather than in isolation, redundant points de-select
-    /// themselves: the second of two interchangeable points has almost nothing left to add.
-    ///
-    /// Cost is `O(count * len())` with a fixed `k` by `k` update, where `k` is at most six.
-    ///
-    /// The running inverse has to start somewhere invertible, so it begins at a small multiple of
-    /// the identity scaled to the data (see `ridge_rel`). This is the standard Bayesian form of
-    /// D-optimality, and its influence is negligible once `k` points have been selected.
-    ///
-    /// Returns indices into the original point set, in the order they were chosen, so truncating
-    /// the result yields the best smaller subset. Fewer than `count` indices come back if the
-    /// candidates run out or if no remaining point can add anything.
-    ///
-    /// # Arguments
-    ///
-    /// * `count`: how many points to choose. Values above the number of available points, or
-    ///   below the free degree-of-freedom count, are permitted but the latter cannot produce a
-    ///   well-conditioned subset.
-    /// * `ridge_rel`: the size of the initial regularizer relative to the mean diagonal of the
-    ///   full information matrix. `None` uses a sensible default; there is rarely a reason to
-    ///   set it.
-    pub fn select_d_optimal(&self, count: usize, ridge_rel: Option<f64>) -> Vec<usize> {
-        let k = self.free.len();
-        if k == 0 || count == 0 || self.rows.is_empty() {
-            return Vec::new();
-        }
-
-        // Scale the ridge to the problem so it stays negligible regardless of the units in play.
-        // A completely empty information matrix (every weight zero) falls back to unity.
-        let scale = self.h.trace() / k as f64;
-        let ridge = ridge_rel.unwrap_or(DEFAULT_RIDGE_REL) * if scale > 0.0 { scale } else { 1.0 };
-
-        // The inverse of `ridge * I` is `I / ridge`.
-        let mut inverse = DMatrix::identity(k, k) / ridge;
-
-        // Precompute the restricted rows once; the inner loop touches all of them every round.
-        let restricted: Vec<DVector<f64>> =
-            self.rows.iter().map(|r| restrict(r, &self.free)).collect();
-
-        let mut chosen = Vec::with_capacity(count.min(self.rows.len()));
-        let mut taken = vec![false; self.rows.len()];
-
-        for _ in 0..count.min(self.rows.len()) {
-            let mut best = None;
-            let mut best_gain = 0.0;
-
-            for i in 0..self.rows.len() {
-                if taken[i] || self.weights[i] <= 0.0 {
-                    continue;
-                }
-                let j = &restricted[i];
-                let gain = self.weights[i] * (j.transpose() * &inverse * j)[(0, 0)];
-                if gain > best_gain {
-                    best_gain = gain;
-                    best = Some(i);
-                }
-            }
-
-            let Some(i) = best else { break };
-
-            // Sherman-Morrison for `H + w j j^T`:
-            //     (H + w j j^T)^-1 = H^-1 - (w H^-1 j j^T H^-1) / (1 + w j^T H^-1 j)
-            let j = &restricted[i];
-            let hj = &inverse * j;
-            let denom = 1.0 + best_gain;
-            inverse -= (&hj * hj.transpose()) * (self.weights[i] / denom);
-
-            taken[i] = true;
-            chosen.push(i);
-        }
-
-        chosen
-    }
-}
-
-/// The indices of the free degrees of freedom, in parameter order.
-fn free_indices(dof: &Dof6) -> Vec<usize> {
-    [dof.tx, dof.ty, dof.tz, dof.rx, dof.ry, dof.rz]
-        .iter()
-        .enumerate()
-        .filter_map(|(i, &free)| free.then_some(i))
-        .collect()
-}
-
-/// Pulls the free components out of a full six-parameter jacobian row.
-fn restrict(row: &AlignStorage3, free: &[usize]) -> DVector<f64> {
-    DVector::from_iterator(free.len(), free.iter().map(|&i| row[i]))
 }
 
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::geom3::align3::AlignSurfMatch3;
+    use crate::geom3::align3::{AlignStorage3, AlignSurfMatch3};
+    use crate::na::Matrix6;
     use crate::{Iso3, Mesh3, UnitVec3, Vector3};
     use approx::assert_relative_eq;
